requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
black==26.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
curl_cffi==0.13.0
fastapi==0.135.1
flake8==7.3.0
frozendict==2.4.7
h11==0.16.0
idna==3.11
isort==8.0.1
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
newsapi-python==0.2.7
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pathspec==1.0.4
peewee==4.0.2
platformdirs==4.9.4
protobuf==7.34.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
yfinance==1.2.0
